pyproject.toml
[project]
name = "cursor-clone-poc"
version = "0.1.0"
description = "PoC agentic coding loop (Plan → Act → Observe) with LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.0",
    "langchain-core>=0.3",
    "langchain-google-genai>=2.0",
    "langchain-openai>=0.2",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
